+++ src/app/types.ts
@@ -1,6 +1,15 @@
+import { ReactNode } from "react"
+
 export interface PageProps {
   params?: {
     slug?: string[] | string
   }
   searchParams?: string[] | string
+}
+
+export interface Route {
+  id: string
+  href: string
+  children: ReactNode
+  categories?: string[]
 }
